+++ core/Updates/4.0.0-b1.php
@@ -9,6 +9,8 @@
 
 namespace Piwik\Updates;
 
+use Piwik\Option;
+use Piwik\Plugins\UserCountry\LocationProvider;
 use Piwik\Updater;
 use Piwik\Updates as PiwikUpdates;
 use Piwik\Updater\Migration\Factory as MigrationFactory;
@@ -30,20 +32,38 @@ class Updates_4_0_0_b1 extends PiwikUpdates
 
     public function getMigrations(Updater $updater)
     {
-        $migration1 = $this->migration->db->changeColumnType('log_action', 'name', 'VARCHAR(4096)');
-        $migration2 = $this->migration->db->changeColumnType('log_conversion', 'url', 'VARCHAR(4096)');
+        $migrations = [];
+        $migrations[] = $this->migration->db->changeColumnType('log_action', 'name', 'VARCHAR(4096)');
+        $migrations[] = $this->migration->db->changeColumnType('log_conversion', 'url', 'VARCHAR(4096)');
 
-        $migration3 = $this->migration->plugin->activate('BulkTracking');
+        $migrations[] = $this->migration->plugin->activate('BulkTracking');
 
-        return array(
-            $migration1,
-            $migration2,
-            $migration3
-        );
+        if ($this->usesGeoIpLegacyLocationProvider()) {
+            // activate GeoIp2 plugin for users still using GeoIp2 Legacy (others might have it disabled on purpose)
+            $migrations[] = $this->migration->plugin->activate('GeoIp2');
+        }
+
+        return $migrations;
     }
 
     public function doUpdate(Updater $updater)
     {
         $updater->executeMigrations(__FILE__, $this->getMigrations($updater));
+
+        if ($this->usesGeoIpLegacyLocationProvider()) {
+            // switch to default provider if GeoIp Legacy was still in use
+            LocationProvider::setCurrentProvider(LocationProvider\DefaultProvider::ID);
+        }
+    }
+
+    protected function usesGeoIpLegacyLocationProvider()
+    {
+        $currentProvider = LocationProvider::getCurrentProviderId();
+
+        return in_array($currentProvider, [
+            'geoip_pecl',
+            'geoip_php',
+            'geoip_serverbased',
+        ]);
     }
 }